### Feature or Bugfix
- Refactor

### Purpose
- see #4803 

### Result

```
# before (HEAD of 1.7)
.tox/py36/bin/sphinx-build -b latex doc build/sphinx/latex  69.88s user 1.18s system 92% cpu 1:16.55 total
.tox/py36/bin/sphinx-build -b latex doc build/sphinx/latex  64.74s user 0.63s system 97% cpu 1:07.04 total
.tox/py36/bin/sphinx-build -b latex doc build/sphinx/latex  64.62s user 0.54s system 98% cpu 1:06.43 total

# after (this PR)
.tox/py36/bin/sphinx-build -b latex doc build/sphinx/latex  18.01s user 0.58s system 89% cpu 20.751 total
.tox/py36/bin/sphinx-build -b latex doc build/sphinx/latex  17.95s user 0.51s system 89% cpu 20.562 total
.tox/py36/bin/sphinx-build -b latex doc build/sphinx/latex  16.42s user 0.42s system 94% cpu 17.761 total
```