@tk0miya 2018-04-14
Note:
In this PR, `sphinx/builders/latex.py` is renamed to `sphinx/builders/latex/__init__.py`.
I think this is not breaking change.