@jfbu 2018-04-15
Even on smaller project than Sphinx own docs I get very significant speed improvement or ``make latex``, for example going down from 10 seconds to 5 seconds for a project using ``latex_show_urls     = 'footnote'`` and ending up with about 120 such footnotes in a 180 pages PDF.

This is very very satisfactory improvement!